extract versioning linters into a separate file
